Fixes a typo and adds a few comma's.
This is the first PR in which I add a few comma's that I think are needed to make the sentence more readable. I haven't done this before, although I think there are a plethora of sentences that would benefit from additional comma's. I just didn't want to come across as a nitpicker.
How do you feel about that?  Shall we leave it at that or should I create a PR for each and every comma I feel is missing? Bear in mind, however, that I am by no means a linguist. Far from it even. It is mostly driven by intuition.